requirements.txt
# Core Framework
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication & Security
bcrypt==4.0.1
python-jose==3.3.0
python-multipart==0.0.6

# Web3 & Blockchain
web3==6.0.0
eth-account==0.8.0

# AI & LLM Integration
langchain==0.0.267
openai==0.28.1
anthropic==0.5.0

# Data Processing
pandas==2.1.1
numpy==1.24.3
python-dotenv==1.0.0

# API Clients
requests==2.31.0
aiohttp==3.8.5

# Database
alembic==1.12.0

# Testing
pytest==7.4.2

# Run with: pip install -r requirements.txt
